separate files under folders and import files in script file
